backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==2.3.3
reportlab==4.4.4
